update master case

+++ behave_dble/features/steps/mysql_steps.py
@@ -196,7 +196,7 @@ def step_impl(context, host_name, num, concur="100", mode_name="user"):
     info_dic = row.as_dict()
     concur = min(int(concur), num)
 
-    tasks_per_thread = num/concur
+    tasks_per_thread = int(num/concur)
     mod_tasks = num%concur
     timestamp = int(round(time.time() * 1000))
 
@@ -206,7 +206,7 @@ def step_impl(context, host_name, num, concur="100", mode_name="user"):
         my_dic["toClose"] = "False"
         last_count = tasks_count-1
         sql_raw = my_dic["sql"]
-        for k in range(tasks_count):
+        for k in range(int(tasks_count)):
             if k==last_count:
                 my_dic["toClose"] = "False"
             id = base_id+k
